backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5
python-dotenv==1.0.1
openai==1.57.4
httpx==0.28.1
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose==3.3.0
sentence-transformers==3.2.1
psycopg2-binary==2.9.10
